iio: adc: qcom-spmi-adc5-gen3: Update register dump print log
Update the register dump print log so that the field width argument
has the correct type.

+++ drivers/iio/adc/qcom-spmi-adc5-gen3.c
@@ -329,7 +329,7 @@ static void adc5_gen3_dump_register(struct regmap *regmap, unsigned int offset)
 			pr_err("debug register dump failed with rc=%d\n", rc);
 			return;
 		}
-		pr_err("%#04x: %*ph\n", offset, sizeof(buf), buf);
+		pr_err("%#04x: %*ph\n", offset, (int)(sizeof(buf)), buf);
 		offset += NUM_BYTES;
 	}
 	pr_err("\n");
